SI-9382 Zippy clean-up in aisle 2 & 3
Consolated JUnit tests and heeded comment about private def and
code beauty.
